+++ src/main/java/com/sparta/fifteen/repository/RefreshTokenRepository.java
@@ -1,0 +1,7 @@
+package com.sparta.fifteen.repository;
+
+import com.sparta.fifteen.entity.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, String> {
+}
