--- src/main/java/com/sparta/fifteen/repository/UserRefreshTokenRepository.java
@@ -1,7 +1,0 @@
-package com.sparta.fifteen.repository;
-
-import com.sparta.fifteen.entity.UserRefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRefreshTokenRepository extends JpaRepository<UserRefreshToken, Long> {
-}
